Correcting

+++ userspace/tinylibc/string/strrchr.c
@@ -2,8 +2,8 @@
 
 char *strrchr(const char *s, int c)
 {
-   size_t len;
-   for (len = strlen(s) - 1; len >=0; len --)
+   long len;
+   for (len = strlen(s) - 1; len >= 0; len --)
       if (s[len] == c) {
          return (char*)&s[len];
       }
